+++ faker/providers/automotive/fr_DZ/__init__.py
@@ -1,0 +1,5 @@
+from ..ar_DZ import Provider as ArDzAutomotiveProvider
+
+
+class Provider(ArDzAutomotiveProvider):
+    pass